Chunk the requests to get_parent_map when finding which revisions touch a file.
bzrlib.log._filter_revisions_touching_file_id suggests that chunking the
requests to get_parent_map is a big win for GraphIndex. We steal the code
from there to do this as it it does some extra things that we don't need.

As the comment in that function says the characteristics will change with
different index formats, so a more robust solutions should be sought.