Initial data load seems to work. Working a little bit on performance.
Baseline time to get the merge_sort graph is around 1.3s (for bzr.dev)
Doing a query-per-revision_id took around 2s to load. Changing that to
batch, dropped it to <300ms.

Makes me wonder if we would be better off doing an INSERT OR IGNORE as the data-load
step, and then have the rest of the inserts be (value = (SELECT id FROM revision WHERE ...)