Distribution: batch size   time 1            2.155 2            1.221 5            0.589 10           0.390 20           0.278 50           0.214 100          0.197 200          0.193 500          0.212 1000         0.220
I think the diminishing returns are probably when the db starts thinking it
needs to do a full table scan. There is also a limiting factor of the
size of 'remaining'.
Anyway, using db ids and using >1 key-at-a-time is a good thing. In the
end, it can even be faster than the known-graph implementation. Though
not by a huge margin.