Support expanding all nodes in the db into their own dotted_revno info.
This is meant to represent the extreme case where every revision is its own branch.
Note that linear revisions are still only 1 branch. However, when importing bzr.dev,
we end up with 4,706 'mainlines', versus ~30,000 revisions.
This is quite slow, and does expand the data a lot.
Specific numbers for bzr.dev:
  Time to just import bzr.dev: 5.642s
  Time to import and expand: 3m22s
  Num expanded nodes: 4706
  Avg time per node: 0.043s

I haven't profiled it, but my guess is that 40ms is the 'merge_sort' time. It is
pretty fast, but obviously not that fast when you are doing 5k of them.

 entries in dotted_revno for bzr.dev: 30,496
 entries in dotted_revno after expansion: 821,488
 avg extra dotted_revno entries per expand: 168

168 new dotted revnos per branch was a bit surprising to me. My guess is that it
primarily stems from all the times when trunk was merged back into dev branches.
(Consider all the NEWS times when you have to merge a lot of stuff back, just to
handle NEWS conflicts.)

I think it does validate the design, at least for bzr. Yes 800k is a lot. But
when you consider that it represents almost 5k branches, versus the idea of
using a 'whole ancestry' table. Which would be far closer to 5k*30k = 141M
It is a savings of 175:1.
Now to try with stuff like mysql...