A couple of fixes to the _find_interesting_ancestry code.
1) Always step the mainline first. We know we need to step it at least once.
Might as well do it first, and get extra filtering.
2) When finished, we may have filtered out ancestors without having
to load their parent info, but we need to load that info in the end.
So add a loop to bring in all LH parent's dotted_revno info.

Now we still need to fix some of the other merge_sort issues, where it
needs to bring in more mainline ancestry to avoid 'race' conditions, etc.
But we are at least back to working order.