Add a helper to the test suite, clean up some code paths.
We were making a mistake when our _split-gdfo call was finding the mainline
id by accident. It was clearly 'not merged' into the mainline, so it was
getting marked interesting, rather than noticing that it *was* the mainline.

Also fixed a bug in the 'create a new branch' code path, which was potentially
an infinite loop.