quite a bit better. Now only 11min (up from 3.5min) to expand all of bzr.dev.
This is down from 20-30min before the recent tweaks.

One is to cache the db_to_rev mapping while we are building the rev<=>db mapping.
Rebuilding that dict on each tip import was too expensive.

Lots of statistics added, to try to help understand what is going on.
Current results for importing bzr.dev:
{'_insert_node_calls': 30531,
 'nodes_expanded': 4713,
 'num_search_tips': 803215,
 'pushed': 791477,
 'pushed uninteresting': 27,
 'ranges_inserted': 52,
 'revs_in_ranges': 5137,
 'split already imported': 30358,
 'split child imported': 9349,
 'split children interesting': 35720,
 'split gdfo': 730354,
 'split parent imported': 6783,
 'step mainline': 3891475,
 'step mainline ensure LH': 32509,
 'step mainline initial': 4713,
 'step mainline null-branch': 2573,
 'step mainline sub-branch': 3202496,
 'step mainline unknown': 649184,
 'step search tips': 82196,
 'total_nodes_inserted': 821981}

lsprof shows that _split_gdfo does a pretty good job and catches quite a few
cases. For bzrtools 'split using children' did a fair share. But for bzr itself
'split using gdfo' seems to do the most of it.
Though also, for bzrtools 'step mainline unknown' was the primary cause of stepping,
but for bzr, 'step mainline sub-branch' takes the cake. Specifically, out of
3.9M mainline steps, 3.2M of them were because we wanted to be able to number
a sub-branch. Good news is that we may be able to improve that by searching
for the 'newest' sub-branch, rather than the 'current' sub-branch.

We also may want to just make step_mainline cheaper by having it step
more than one revision at a time. Possibly using gdfo as a hint for how far to go...

Note that walk_ancestry --db-db-id was 300+ms, but --db-dotted was only 87ms.