map(int, revno_str.split('.')) shows up high on lsprof.
We may want to switch to storing based on integer revno bits, and use
either (0, 0, X) for the mainline or (-1, -1, X).

However, our best bet is to avoid walking the mainline so much, right
now, having to step-back 3.9M times (~3.2M sub-branch, 650k for unknown)
is causing us to load 20M dotted_revno entries.

The other possibility is to cache the mainline => dotted_revno data,
which should help, but I'd like to avoid it for now.