Another dramatic improvement. Now *almost* as fast as the old code :).
Specifically, instead of stepping all the way back to X.Y.1, we step back until we
find *any* X.?.1 revision. Once we've found one, we know that we have the latest.

Now down to 4m3s, which is close to the ~3m30s that I think we used to have.
Instead of bringing in 20M revisions, we now bring in 5M while walking mainline.
We still have 650k mainline steps fro 'unknown' but now we are way down to
only 113k because of sub-branches. *Huge* improvement (was 3M+).
split_gdfo still filters more than split children.

My guess is that we have a lot more branches that are merged multiple times, and
so there are extra 'children' that are actually merged later than our current tip.
However, because they are based on the current revision, their gdfo isn't actually
high enough to be quickly filtered.

Next possible steps-
  mainline rev cache, make it cheaper to step since we're doing it repeatedly
  step by more-than-one-at-a-time. Using the mainline info we already store.
  If we are *just* looking at the branch_count, we could use a new sql query
    SELECT * FROM dotted_revno WHERE ... AND revno LIKE X.?.1
  So we walk the mainline, and try to quickly load the most recent revno.