turns out you also have to handle the reverse mapping.
At this point, even without 'iter_merge_sorted' we show an improvement for
'bzr log -r X.Y.Z..X.Y.Z+5', as it tries to walk them as a simple 'mainline'
first, but does a revision_id => revno lookup at each step.

We need to consider caching something on the Branch object, so that we
don't have to do as much config overhead while processing.