iter_merge_sorted_revisions now basically works.
I don't think it is perfect yet, as we don't handle the various flags
that can be supplied for stop_rule.
However, time for 'bzr log -n0 -r -10..-1' or -r-100..-1 goes from
1.4s down to 0.97s.
And so far the result is 'correct'. I assume because the default
is 'with-merges', which should match the 'skip to lh parent and
treat lh parent as a stop rev.'
At least as long as both parts of the range are mainline revs.