Fix a bug in dotted => revision_id.
We were failing to terminate the loop once all revno_strs had been found.
So we always walked the full ancestry. Both because we were discarding
the wrong object, and because we weren't checking that there were
no more objects to search for.

For emacs, this makes a lookup for the most recent rev take 1ms, instead of 500ms.