Iron out some bugs.
It now auto-imports on demand for any of dotted=>rev rev=>dotted and iter_merge.
It can be a bit slow, but not terrible. And it means it is always correct
and the next log -n0 should be faster.
It does slow down log -n1 -r-10..-1 slightly, 700ms=>750ms or so. I'm not
sure if that really matters.