A couple small tweaks.
Seems to be about 10% improvement for emacs, not noticeable for bzr.dev.
Don't build the dotted revno cache if we won't be using it.
create a local var for a function, use a local function for profiling.
Use the db id since we already know it, rather than joining against revision
all the time.