Down to 17.9s
Get rid of a LBYL section. We don't need to check if a node is already
imported, we'll just get a UNIQUE constraint error when we try to insert it.
If we are doing incremental, the merge_sorted result won't include extra
nodes anyway, so we don't need to check 2 times.