Set VIRTUALENV_DOWNLOAD=0.
This is less friendly to running tests on focal, where virtualenv seems to
end up with a too-new pip for Python 3.5 even downloading is disabled via
its app-data directory; but it works on xenial, where tox doesn't support
configuring virtualenv's downloading behaviour flexibly enough.  I haven't
been able to find a way for both old and new systems to work properly.
Hopefully by the time this matters very much we'll be able to drop support
for 3.5.