Port from SimpleTAL to Chameleon.
SimpleTAL is a painful dependency, because it has no download files on
PyPI: the only way to depend on it properly is therefore using a direct
URL reference, and as I understand it packages with such references
those can't be uploaded to PyPI.

Porting to Chameleon (Zope's default TAL implementation nowadays) turns
out not to be too difficult.  The main work involved is because it
doesn't support path expressions.  It would be possible to add those
using `z3c.pt`, and I considered that, but that has a significant
dependency tree for packages that don't already use the Zope Toolkit,
and it would have also required extra startup complexity to register
`ITraversable` adapters.  It's easier to convert path expressions to the
equivalent Python expressions, although the process of doing so was
unfortunately quite error-prone.